Distinguish has-enrolled from has-exited in `is_enrolled`
Is this the best solution? probably not, but I'm not sure about what the
best is right now. I'd rather get it in front of people first.

I do believe that it is useful for project staff to quickly see which
households haven't yet been enrolled vs those that have already been
exited.