Add filter for assessment completion status.
Values are: 0 (incomplete) or 1 (up to date). Incomplete means that
there is an entry date but no entry assessment, or that there is an
exit date but no exit assessment. Up to date means that all members
that have an entry date also have an entry assessment, and that all
members that have an exit date also have an exit assessment.